requirements.txt
aiocontextvars==0.2.2
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.1
billiard==3.3.0.23
black==20.8b1
botocore==1.20.23
celery==3.1.18
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
contextvars==2.4
cryptography==3.4.6
dataclasses==0.8
decorator==4.4.2
Django==3.1.7
djangorestframework==3.12.2
dynaconf==3.1.4
et-xmlfile==1.0.1
hvac==0.10.8
idna==2.10
immutables==0.15
jdcal==1.4.1
jmespath==0.10.0
kombu==3.0.37
loguru==0.5.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
openpyxl==3.0.6
pandas==1.1.5
pathspec==0.8.1
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
redis==2.10.3
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
semantic-version==2.8.5
setuptools-rust==0.12.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
validators==0.18.2
